Removed bias from map generator
Previously, the map generation algorithm seemed to be very slightly
biased in favour of presenting the Nth item in the mini test. This has
been corrected by reshuffling the N items in each mini set.
Also made feedback_time a global parameter and the colour of the words
now change to green or red in feedback depending on whether or not they
were correct.