Now returns natual log of dT
dT is now defined at the natural log of 1 + the sum of the Euclidean
distances between points. This log transform avoids the problem where
scaling super-skinny triangles results in a triangle that is crazy long
and therefore extremely distant from another more normal triangle
(which in turn skews the Pearson correlation).