Fixed rotation bug
If a triangle is already rotated exactly north or exactly south,
rounding error can cause arccos to fail and return nan. Added a
workaround to detect this: if the triangle is already pointing north,
the function returns the triangle untouched; if the triangle is
pointing south, theta is set to pi before doing the rotation.