Modifications to run many chains or metrics
Added new code to run all metrics for all chains. Modified code to
handle the measurement of structure in the dynamic and combined sets.
Also, pre-computing the meaning distances for the stable set, so that
the distances aren't repeatedly computed. plotAll now plots both the
positive and negative 95% confidence level.