Added code for shape-size scatter plots
Added code to produce shape-size scatter plots, which are 2-dimensional
plots of shape vs. size. Triangles are then plotted in the space and
colour-coded by word. If there are a large number of different words,
it becomes impractical to visualize the language this way, so
hierarchical clustering is used to turn a large number of words into a
smaller set of clusterings based on normalized Levenshtein distance.