Final changes
This commit includes all the final changes that were made prior to
submission of the dissertation. The main additions are:

• Added option to runStats() to include/ignore gen-0 data points.

• Fixed calculation of entropy and conditional entropy. Also added the
option to include a start/stop character.

• Fixed intergenCorr() function to calcualte the partial correlation
between structure and learnability.

• Changed triangleOverlayGraphic() function so that it outputs an SVG
file rather than HTML so that you can easily put vector graphics into
your document.

• Added drawTriangles() function so that you can draw an arbitrary set
of triangles and write out as SVG.

• Added wordMemory() function which calculates the "memory performance"
of participants. This measure was not included in the final version of
the dissertation to save space, but essentially the idea is to quantify
how many words the participant correctly remembered from their training
material regardless of whether they associated them with the correct
meanings.

• Added soundSymbolism() function to calculate the mean pointedness of
triangles associated with different sounds.