Corrected error in Mantel test
My previous method shuffled the distance vector rather than the
distance matrix for the sake of efficiency. It turns out that this is
wrong and exaggerates the Z-scores. The issue is corrected in this
commit. Added a separate Mantel test module (MantelTest.py). For more
info, see my blog post here:
http://www.jonwcarr.net/blog/2014/9/19/a-guide-to-the-mantel-test-for-li
nguists