Changed percentage score
to a points-based score. You now get 10 points for every successful
match. Percentages are not a good way to score this game because: (i)
towards the end of the experiment, percentages only increase by a small
amount, and (ii) 200 points feels much more encouraging that 20% even
through the two are approximately equivalent by the end of the
experiment.